Milk-Tracking-App/agent/pyproject.toml
[project]
name = "milk-tracking-agent"
version = "0.1.0"
description = "AI-powered Milk Tracking Agent using OpenAI Agents SDK with Gemini"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.1.0",
    "openai>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "chainlit>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
